+++ bench/workloads/dependency-graph/b.js
@@ -1,3 +1,0 @@
-// TODO: depends on a.js. Export the number a.x + 2.
-const a = require('./a.js');
-module.exports = null;