+++ bench/workloads/dependency-graph/c.js
@@ -1,3 +1,0 @@
-// TODO: depends on a.js. Export the number a.x * 5.
-const a = require('./a.js');
-module.exports = null;