+++ bench/workloads/dependency-graph/d.js
@@ -1,4 +1,0 @@
-// TODO: depends on b.js AND c.js. Export the number b + c (which should be 8).
-const b = require('./b.js');
-const c = require('./c.js');
-module.exports = null;